requirements.txt
fastapi
uvicorn
faiss-cpu
sentence-transformers
pymupdf
beautifulsoup4
requests
ollama